:white_check_mark: Add test case for AutoComplete
